prepare for duel
