+++ duel/src/main/java/me/hsgamer/flexegames/game/duel/DuelGame.java
@@ -1,23 +1,24 @@
 package me.hsgamer.flexegames.game.duel;
 
 import me.hsgamer.flexegames.api.game.Game;
+import me.hsgamer.flexegames.api.property.GamePropertyMap;
 import me.hsgamer.flexegames.builder.ItemBuilder;
 import me.hsgamer.flexegames.util.ItemUtil;
 import me.hsgamer.minigamecore.base.Arena;
 import me.hsgamer.minigamecore.base.ArenaManager;
 import net.kyori.adventure.text.Component;
+import net.minestom.server.entity.Player;
 import net.minestom.server.item.ItemStack;
 
 import java.util.List;
-import java.util.function.Function;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
 
 public class DuelGame implements Game {
     private final DuelExtension duelExtension;
-    private final DuelGameConfig gameConfig;
 
-    public DuelGame(DuelExtension duelExtension, DuelGameConfig gameConfig) {
+    public DuelGame(DuelExtension duelExtension) {
         this.duelExtension = duelExtension;
-        this.gameConfig = gameConfig;
     }
 
     @Override
@@ -27,29 +28,30 @@ public class DuelGame implements Game {
 
     @Override
     public Component getDisplayName() {
-        return gameConfig.getDisplayName();
+        return duelExtension.getMessageConfig().getDisplayName();
     }
 
     @Override
     public List<Component> getDescription() {
-        return gameConfig.getDescription();
+        return duelExtension.getMessageConfig().getDescription();
     }
 
     @Override
     public ItemStack getDisplayItem() {
         return ItemUtil.stripItalics(
-                ItemBuilder.buildItem(gameConfig.getDisplayItem())
+                ItemBuilder.buildItem(duelExtension.getMessageConfig().getDisplayItem())
                         .withDisplayName(getDisplayName())
                         .withLore(getDescription())
         );
     }
 
     @Override
-    public Function<ArenaManager, Arena> createArena(String name) {
-        return arenaManager -> new DuelArena(duelExtension, this, name, arenaManager);
+    public CompletableFuture<GamePropertyMap> editProperty(Player player, GamePropertyMap gamePropertyMap) {
+        return CompletableFuture.completedFuture(gamePropertyMap);
     }
 
-    public DuelGameConfig getGameConfig() {
-        return gameConfig;
+    @Override
+    public Arena create(String name, GamePropertyMap gamePropertyMap, ArenaManager arenaManager, UUID owner) {
+        return new DuelArena(duelExtension, name, gamePropertyMap, this, arenaManager, owner);
     }
 }
