--- duel/src/main/java/me/hsgamer/flexegames/game/duel/feature/ConfigFeature.java
@@ -1,7 +1,0 @@
-package me.hsgamer.flexegames.game.duel.feature;
-
-import me.hsgamer.flexegames.game.duel.DuelGameConfig;
-import me.hsgamer.minigamecore.base.Feature;
-
-public record ConfigFeature(DuelGameConfig config) implements Feature {
-}
