+++ duel/src/main/java/me/hsgamer/flexegames/game/duel/feature/GameDescriptionFeature.java
@@ -2,6 +2,7 @@ package me.hsgamer.flexegames.game.duel.feature;
 
 import me.hsgamer.flexegames.builder.ItemBuilder;
 import me.hsgamer.flexegames.feature.arena.DescriptionFeature;
+import me.hsgamer.flexegames.game.duel.DuelExtension;
 import me.hsgamer.flexegames.util.ItemUtil;
 import me.hsgamer.flexegames.util.TimeUtil;
 import me.hsgamer.minigamecore.base.Arena;
@@ -18,9 +19,11 @@ import java.util.function.Supplier;
 
 public class GameDescriptionFeature implements DescriptionFeature {
     private final Arena arena;
+    private final DuelExtension duelExtension;
 
-    public GameDescriptionFeature(Arena arena) {
+    public GameDescriptionFeature(Arena arena, DuelExtension duelExtension) {
         this.arena = arena;
+        this.duelExtension = duelExtension;
     }
 
     @Override
@@ -37,7 +40,7 @@ public class GameDescriptionFeature implements DescriptionFeature {
     @Override
     public ItemStack getDisplayItem() {
         return ItemUtil.stripItalics(
-                ItemBuilder.buildItem(arena.getFeature(ConfigFeature.class).config().getArenaDisplayItem(), getReplacements())
+                ItemBuilder.buildItem(duelExtension.getMessageConfig().getArenaDisplayItem(), getReplacements())
         );
     }
 }
