+++ duel/src/main/java/me/hsgamer/flexegames/game/duel/state/EndingState.java
@@ -3,7 +3,6 @@ package me.hsgamer.flexegames.game.duel.state;
 import me.hsgamer.flexegames.api.game.ComponentDisplayName;
 import me.hsgamer.flexegames.feature.arena.DescriptionFeature;
 import me.hsgamer.flexegames.game.duel.DuelExtension;
-import me.hsgamer.flexegames.game.duel.feature.ConfigFeature;
 import me.hsgamer.flexegames.game.duel.feature.InstanceFeature;
 import me.hsgamer.flexegames.game.duel.feature.WinnerFeature;
 import me.hsgamer.flexegames.manager.ReplacementManager;
@@ -24,15 +23,14 @@ public class EndingState implements GameState, ComponentDisplayName {
 
     @Override
     public void start(Arena arena) {
-        var gameConfig = arena.getFeature(ConfigFeature.class).config();
         var winnerFeature = arena.getFeature(WinnerFeature.class);
         var descriptionFeature = arena.getFeature(DescriptionFeature.class);
         arena.getFeature(TimerFeature.class).setDuration(gameConfig.getEndingTime(), TimeUnit.SECONDS);
         Component message;
         if (winnerFeature.getWinner() != null) {
-            message = ReplacementManager.replace(gameConfig.getWinnerMessage(), descriptionFeature.getReplacements());
+            message = ReplacementManager.replace(duelExtension.getMessageConfig().getWinnerMessage(), descriptionFeature.getReplacements());
         } else {
-            message = gameConfig.getNoWinnerMessage();
+            message = ReplacementManager.replace(duelExtension.getMessageConfig().getNoWinnerMessage(), descriptionFeature.getReplacements());
         }
         arena.getFeature(InstanceFeature.class).clearInventory();
         arena.getFeature(InstanceFeature.class).sendMessage(message);
